Refactor test to use remote file and temporary workspace.
